o removed KileCapability tester classes, didn't work properly and should be replaced with something more general anyway
o granted wish 64856: More advanced build system, Kile is now able to autodetect whether or not BibTeX, MakeIndex should be run, also LaTeX is executed until the "Rerun" warnings are gone.

o Add the bibliography as a dependency for the TeX file, and also add it to the structure view.