Various bugfixes by Holger Danielsson: 1) the member 'm_boundingbox' was not set with the current entry from the config file. So it was 'false' and no bounding box information was inserted.
2) The filter string for choosing an image in the KFileDialog was
wrong, because this must be a space separated list of shell globs
(not comma separated). So no images were shown under 'Graphics'.

3) The regexp for the recognizing the resolution is no longer correct.
Newer versions of imagemagick will also use floating point numbers,
not only integers.

4) Most important: imagemagick now works with PixelsPerInch and
PixelsPerCentimeter. Beginning with version 6.0 of imagemagick, all
my images were too great. So I changed the calculation of the
bounding box: if PixelsPerCentimeters is recognized by the script,
the bounding box is scaled down with factor 2.54. But the resolution
ist multiplied with 2.54, because it is shown als PixelsPerInch in
the includeGraphics dialog.

CCMAIL:holger.danielsson@t-online.de