BUG: 164231 BUG: 174003
Again some codecompletion fun :)
This bug was hidden quite deeply but due to the help of  
Russell W. Howell <howell@*> and one one my flatmates the error could be spotted and solved.
Thanks!

This patch should be tested some more, although it didn't show up any drawbacks at our testing.

the patch:
the error is that if m_ref is true but the m_inprogress false the internal state was inconsistent and therefore the
completion list didn't show up anymore after inserting letters. This is fixed now.

Happy Testing!