don't create shortcuts in the tabBar
this was a hard one, no documentation was useful neither a look at the
source code.
After a long time I found http://bugs.kde.org/121782 which mentioned
setNoAccel() from KAcceleratorManager and this was the solution.
Heureka!

the other change avoids setting the tabText twice and now we always use
documentName as source for the tab text.