rewrite action handling to not use actionLists
This is needed due to bug #194732. The interesting thing is that now
kile seems a bit faster, at least on my machine.

This patch makes it now possible to change any shortcuts you want and
they will be preserved upon restarts.
One not so nice thing is that the shortcuts for tools and user tags are
not the default ones so restoring the default shortcuts removes these
shortcuts completely.

CCBUG: 187965