Don't block signals from the file browser widget.
This was introduced to avoid opening a single file twice by double-clicking on it in the file browser widget. Due to a bug in that code, the blocking of the signals wasn't released in some circumstances. Currently, blocking signals is no longer necessary as 'fileOpen' can't be called when 'fileOpen' is being executed, which was possible due to the event processing during parsing.

BUG: 222325