revive "Auto insert $"
"Auto insert $" got lost during the KDE3 -> KDE4 migration. This patch
implements the auto-dollar feature just like in the KDE3.x days.
If the user types '$', the completion box is always hidden because there
is no match. Hence, right now it is not checked whether the completion
is active (see comment in the code)

CCMAIL: klistman@rtr.tu-darmstadt.de