Avoid crashes when running Kile in a non-English locale.
In the EditorExtension class it was assumed that the opening and closing
variants for double quotes in a given option string for inserting double quotes
are always separated by two space characters. Unfortunately, this option string
was also passed to an i18n function call. Hence, in some languages (e.g. French)
these two spaces were sometimes not added to the translation, which could result
in a crash when the opening and closing quotes were to be extracted from the
option string. This has now been fixed by using two lists for internal and
external (i.e. internationalised) double quote options.

BUG: 271910