Avoid crashes due to an event loop being executed while code in some tool class is running
It must be ensured that no event loop is started whenever some tool operation is running. This
includes running code inside the tool classes!

The reason for that is that an event loop might trigger the deletion a tool object for which code
is currently executed, for instance with the 'stopLivePreview' method.
An event loop is executed, for example, within the 'documentSave' method of KatePart. Although the event
loop doesn't process user events, the document modification timer might still be triggered in such
an event loop and 'stopLivePreview' will be called. Now, no document saving is performed inside tool
classes anymore (including the tool manager).