Associate the (LaTeX) compilation output with the document/projects it originates from.
In that way the correct output can be restored when switching between documents.
Also, add a tab widget to the output pane which allows to choose between displaying
error, warning or badbox messages, and add a label to the status bar which
informs about the number of errors, warnings or badboxes.