Add quick access menu for open documents.
The patch does two things:
1. replace QTabWidget by separate tabbar:
Simplifies structure of used views and class variables. All texteditor views
are now available via the tab bar's tabData(index) of the respective tabs.

2. add a quick access menu:
Provide a menu for all open documents in a ordered list. This is especially usefull if
having a long list of open documents.

REVIEW: 124367