Remove Tip of the Day.
Support for this feature is already removed in most of all KDE applications.
Visual design people strongly decourage using it, mainly for the reason
that the interfaces must be self-explaining enough and let the user
explore for interesting features. More advanced features have to be
explained in the docbook.

REVIEW: 124538