Fix location of temporary lyx pipes
Make sure that there is a directory separator when creating the path
of the temporary lyx pipes.  This changes what is created in the
temporary directory: instead of
├── [drwx------]  kile-XXXXXX
├── [drwx------]  kile-XXXXXX.lyx
│   ├── [prw-------]  lyxpipe.in
│   └── [prw-------]  lyxpipe.out
├── [prw-------]  kile-XXXXXX.lyxpipe.in
└── [prw-------]  kile-XXXXXX.lyxpipe.out
now is it
└── [drwx------]  kile-XXXXXX
    ├── [drwx------]  .lyx
    │   ├── [prw-------]  lyxpipe.in
    │   └── [prw-------]  lyxpipe.out
    ├── [prw-------]  .lyxpipe.in
    └── [prw-------]  .lyxpipe.out
which also leaves no empty kile-XXXXXX.lyx directories at the exit.

This bug was introduced when porting from KTempDir to QTemporaryDir:
KTempDir::name() returns the full path with a trailing slash, while
QTemporaryDir::path() does not have the trailing slash.