Fix frame title in structure view
The frame may have additional option '[fragile]' used when verbatim or similar kind of environments are used in it. Then, the frame title is not recognised as parser looks directly for title inside {} brackets. This commit fixes this and the frame name is properly displayed in the structure tree.

Also improve whitespace handling for 'block', 'exampleblock', and 'alertblock' titles.