Revert "Adjust URL for help in appdata XML"
The change is not going to fix the issue referenced.
We should not refer directly to full path, because
that path is meant to be changed and dynamically solved
by docs.kde.org.
If there is a global issue with the resolution, that would
affect all programs, so let's fix it globally.
But the change being reverted is not in line with the pattern
that should be used to reference the docs.kde.org URLs.

This reverts commit e90181c6c933431f7fe635d3446fbf2103a822d2.