Add "Open Containing Folder" action
This commit adds an action to the File menu which will open the
default file manager and highlight the current file there.

![grafik](/uploads/30e29419f636be154f58a78398f6ee18/grafik.png)

Many KDE applications (Kate, kdenlive, KFind, Okular, …) have such
an action already because it is very handy in various workflows.
It makes sense to add it based on consistency alone, however I also
personally miss it every time I use Kile because I tend to manually
backup the file I just edited when I am done, but for this I need
to find the file in the Dolphin file manager first. This procedure
is greatly simplified by adding this action.

-----

Hi, nice to meet you. This is my first Kile MR.