Find bibliography tools using base names
This fixes one of the bugs highlighted in https://bugs.kde.org/show_bug.cgi?id=408890, namely that Biber (or any other bibliography backend) can't be configured with an absolute path if desired to be picked up by the automatic detection. The issue stems from comparing the user's input (e.g., `/usr/bin/biber`) to the backend name (e.g., `Biber`) case-insentively, but literally. Changing this to use the base name of the tool command instead resolves the issue.

CCBUG: 408890