Remove requirement of non-empty values in config
While playing with settings, I found out that values of the tool config fields cannot be empty.
If the config value was non-empty before, clearing the config value in the dialog has no effect on save.

With this commit I added a list of fields allowed to be empty on save.

I wasn't sure whether "options" should be there. All the default options have some parameters, even if this is only `%placeholder`, however I can imagine a tool which do not take any value so I think "options" should also be allowed to be empty.

One has to check whether there are other config keys allowed to be empty not included yet.